Updated
Too much to say to be honest

+++ Buffle/Methods.py
@@ -1,133 +1,85 @@
 import os
 import shutil
 
+
 # Methods will not be accessed from this file but from Buffle itself
 # These are to be used from the player from the parent folder Buffle and to be use dy other .py files
 
 # region DISPLAY_RESULTS
-# [0] display_file_results
-# [1] display_text_results
-# [2] display_texture_results
-# [3] display_manual_results
-# enabled or disabled by user to display relevant information
-__displays_results = [True, True, True, True]
-# formats for how the file location is displayed
-# 0 = full: full file directory
-# 1 = limited: 2 directories deep
-# 2 = file: only file
-__file_format = 0
+class DisplayResults:
+    # formats for how the file location is displayed
+    # 0 = full: full file directory
+    # 1 = limited: 2 directories deep
+    # 2 = file: only file
+    __file_format = None
+    # enabled or disabled displaying relevant information
+    __display_results = None
 
 
-def display_file_format(format: int = None):
-    """
-    The format used when displaying file's directory / location
-    0 = full | 1 = limited | 2 = file
-    :type format: int
-    :return Format of displayed file's directory / location
-    """
-    if format is not None:
-        global __file_format
-        __file_format = format
-    return __file_format
+    def __init__(self, file_format: int, display_results: bool):
+        self.file_format = file_format
+        self.display_results = display_results
 
 
-def display_all_results(display: bool = None):
-    """
-    Enables or Disables displaying results from all file & folder altering methods
-    :type display: bool
-    :return If results are enabled or disabled for all file & folder altering methods
-    """
-    if display is not None:
-        for dr in __displays_results:
-            dr = display
-    return __displays_results
+    @classmethod
+    def set_format(cls, format: int = None):
+        """
+        The format used when displaying file's directory / location
+        0 = full | 1 = limited | 2 = file
+        :type format: int
+        :return Format of displayed file's directory / location
+        """
+        if format is not None:
+            cls.__file_format = format
+        return cls.__file_format
 
 
-def display_file_results(display: bool = None):
-    """
-    Enables or Disables displaying results from File methods
-    :type display: bool
-    :return If results are enabled or disabled
-    """
-    if display is not None:
-        __displays_results[0] = display
-    return __displays_results[0]
+    @classmethod
+    def set_display(cls, display: bool = None):
+        """
+        Enables or Disables displaying results from this package's altering methods
+        :type display: bool
+        :return If results are enabled or disabled for this package's altering methods
+        """
+        if display is not None:
+            cls.__display_results = display
+        return cls.__display_results
 
 
-def display_text_results(display: bool = None):
-    """
-    Enables or Disables displaying results from Text methods
-    :type display: bool
-    :return If results are enabled or disabled
-    """
-    if display is not None:
-        __displays_results[1] = display
-    return __displays_results[1]
+    @classmethod
+    def result(cls, source: str, method: str, updated_value, original_value):
+        """
+        Displays the output and processes of a method altering files
+        :param source: Target file's or folder's directory / location
+        :type source: str
+        :param method: name of the method / action done
+        :type method: str
+        :param updated_value: value after alteration
+        :param original_value: value before alteration
+        """
+        try:
+            if os.path.isfile(source) or os.path.isdir(source):
+                if cls.__file_format == 1:  # limited
+                    limited_parts = source.split(os.sep)[-3:]
+                    source = os.path.join(*limited_parts)
+                elif cls.__file_format == 2:  # file
+                    source = source.split("\\")[-1]
 
-
-def display_texture_results(display: bool = None):
-    """
-    Enables or Disables displaying results from Texture methods
-    :type display: bool
-    :return If results are enabled or disabled
-    """
-    if display is not None:
-        __displays_results[2] = display
-    return __displays_results[2]
-
-
-def display_manual_results(display: bool = None):  # not implemented yet or will be till future updates
-    """
-    Enables or Disables displaying results from all Manual methods
-    :type display: bool
-    :return If results are enabled or disabled
-    """
-    if display is not None:
-        __displays_results[3] = display
-    return __displays_results[3]
-
-
-def result(file: str, method: str, updated_value, original_value):
-    """
-    Displays the output and processes of a method altering files
-    :param file: Target file's directory / location
-    :type file: str
-    :param method: name of the method / action done
-    :type method: str
-    :param updated_value: value after alteration
-    :param original_value: value before alteration
-    """
-    try:
-        if os.path.isfile(file):
-            if __file_format == 1:  # limited
-                limited_parts = file.split(os.sep)[-2:]
-                file = os.path.join(*limited_parts)
-            elif __file_format == 2:  # file
-                file = file.split("\\")[-1]
-
-            if original_value != updated_value:
-                print(f"{file} <|> {method} <|> altered <|> [{original_value}] -> [{updated_value}]")
+                if original_value != updated_value:
+                    print(f"{source} <|> {method} <|>   altered <|> [{original_value}] --> [{updated_value}]")
+                else:
+                    print(f"{source} <|> {method} <|> unaltered <|> [{updated_value}]")
             else:
-                print(f"{file} <|> {method} <|> unaltered <|> [{updated_value}]")
-        else:
-            print(f"{file} <|> ERROR <|> file not found")
-    except:
-        print(f"{file} <|> ERROR <|> displaying alters")
+                print(f"{source} <|> ERROR <|> not file or folder")
+        except:
+            print(f"{source} <|> ERROR <|> displaying alters")
 
 
-def __results(file: str, method: str, updated_value, original_value):
-    try:
-        if os.path.isfile(file):
-            if original_value != updated_value:
-                print(f"{file} <|> {method} <|> altered <|> [{original_value}] -> [{updated_value}]")
-            else:
-                print(f"{file} <|> {method} <|> unaltered <|> [{updated_value}]")
-        else:
-            print(f"{file} <|> ERROR <|> file not found")
-    except:
-        print(f"{file} <|> ERROR <|> displaying alters")
-
-
+# all the DisplayResults classes
+display_texture_results = DisplayResults(1, True)
+display_text_results = DisplayResults(1, True)
+display_file_results = DisplayResults(1, True)
+display_manual_results = DisplayResults(1, True)
 # endregion
 
 
@@ -139,24 +91,32 @@ def dump(source: str, delete: bool = False):
     :type source: str
     :param delete: Deletes the target folder's directory / location after dumped
     :type delete: bool
+    :return destination
     """
-    move(source, os.path.dirname(source), delete)
+    return move(source, os.path.dirname(source), delete)
 
 
 def move(source: str, destination: str, delete: bool = False):
     """
-    Moves files from one folder to another folder
+    Moves files from one folder or a single file to another folder
     :param source: File's old directory / location
     :type source: str
     :param destination: File's new directory / location
     :type destination: str
     :param delete: Deletes the source folder's directory / location after moved
     :type delete: bool
+    :return destination
     """
-    for file in os.scandir(source):
-        shutil.move(file, destination)
+    if source is os.path.isdir():  # multiple files in a folder
+        for file in os.scandir(source):
+            shutil.move(file, destination)
 
-    os.scandir(source)
-    if delete:
-        os.remove(source)
+        os.scandir(source)
+        if delete:
+            os.remove(source)
+
+    if source is os.path.isfile():  # single file in a folder
+        shutil.move(source, destination)
+
+    return destination
 # endregion
